+++ src/lib.rs
@@ -57,13 +57,13 @@ extern crate syn;
 #[macro_use]
 mod macros;
 
-mod expr;
 mod item;
+mod pat;
 mod path;
 mod print;
 
-pub use self::expr::*;
 pub use self::item::*;
+pub use self::pat::*;
 
 use proc_macro2::TokenStream;
 use syn::punctuated::Punctuated;
